refactor: 채팅 log 컴포넌트 구조 변경
